requirements.txt
wheel==0.40.0
openai==0.27.4
tiktoken==0.3.3
chromadb==0.3.21
faiss-cpu==1.7.3
hydra-core==1.3.2
langchain==0.0.141
unstructured==0.5.12
huggingface-hub==0.13.4